// src/app/api/models/user.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface User {
  business_address?: string | null;
  business_name?: string | null;
  created_at?: string;
  default_currency?: string | null;
  email?: string;
  id?: string;
  logo_url?: string | null;
}